Picking pagination from the options

+++ recommendation/engine/pagination.go
@@ -1,6 +1,6 @@
 package engine
 
 type Pagination struct {
-	StartIndex    int64
-	NumberOfItems int64
+	StartIndex    int
+	NumberOfItems int
 }
